add backend Dockerfile, so the image can be built without backend's vendor
Previously, to build the Laravel sail runtime image for the backend we needed to
install the application dependencies via composer. After  that,  the  Dockerfile
would be avaiable at vendor/laravel/sail/runtimes.

However,  this  means  need  `composer`  installed  before  even  creating   the
environment in which the backend runs. To avoid it, we store the Dockerfile as a
part of the git repository.