make the frontend container an external one
The paywall now proxies the request to an external docker container
defined by the `FRONTEND` variable. This variable should specify the
name of the container or IP and the port.

This will be possible by connecting the two containers via a common
docker network defined by `FRONTEND_NETWORK`. The frontend container
must be attached to this network beforehand.