feat(session): add expire and cache-control http headers
This prevents browse caching from loading cached assets after the user's
session has expired.

We are awaer we cannot prevent the user from caching  assets  previously
sent to him. But that is not a problem because the point of this app  is
to be a paywall for on-demand web services. Web  services  that  provide
static content won't benefit from this app, but those providing  dynamic
content (such as streaming, gaming, and so on) will.